pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f1sim"
version = "0.0.1"
description = "Modular F1 Race Prediction System (MVP scaffold)"
authors = [{ name = "Waqas Ahmed" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"pandas>=2.2",
	"numpy>=1.26",
	"scikit-learn>=1.4",
	"xgboost>=2.0",
	"catboost>=1.2",
	"fastf1>=3.3",
	"typer>=0.12",
	"rich>=13.7",
	"pydantic>=2.7",
	"pyyaml>=6.0",
	"orjson>=3.10",
	"sqlalchemy>=2.0",
	"duckdb>=1.0",
	"pyarrow>=15",
	"joblib>=1.4"
]

[project.optional-dependencies]
dev = [
	"ipykernel",
	"jupyter",
	"black",
	"isort",
	"ruff",
	"pytest",
	"pytest-cov"
]

[project.scripts]
f1sim = "f1sim.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100 